+++ sharedUiComponents/lines/valueLineComponent.tsx
